PPSS/Prácticas/P05-Dependencias2/P05-mocks/gestorLlamadasMocks/src/main/java/ppss/GestorLlamadas.java
package ppss;

//paquete ppss
public class GestorLlamadas {
    static double TARIFA_NOCTURNA=10.5;
    static double TARIFA_DIURNA=20.8;

    public Calendario getCalendario() {
        Calendario c = new Calendario();
        return c;
    }

    public double calculaConsumo(int minutos) {
        Calendario c = getCalendario(); // dep1
        int hora = c.getHoraActual(); // dep2
        if(hora < 8 || hora > 20) {
            return minutos * TARIFA_NOCTURNA;
        } else {
            return minutos * TARIFA_DIURNA;
        }
    }
}
